Issue fixed.

+++ index.js
@@ -1,6 +1,6 @@
 const PI = Math.PI;
 const userInputRadius = prompt("Please enter radius");
-const radiusNumber = parseInt(userInputRadius);
+const radiusNumber = parseFloat(userInputRadius);
 const area = PI*(Math.pow(radiusNumber, 2));
 const perimetr = 2*PI*radiusNumber;
 console.log(area, 'area');
